Move all layout components into dedicated /layouts folder
