+++ frontend/src/grids/index.ts
@@ -1,7 +1,4 @@
 // Grids barrel file
-export { default as HomeLayout } from './HomeLayout';
-export { default as IdeaLayout } from './IdeaLayout';
-export { default as UserLayout } from './UserLayout';
 export { default as IdeaGrid } from './IdeaGrid';
 export { default as InfoGrid } from './InfoGrid';
 export { default as UserGrid } from './UserGrid';
