+++ frontend/src/layouts/HomeLayout.tsx
@@ -3,11 +3,12 @@ import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
 import { Styles } from 'jss';
 import Grid from '@material-ui/core/Grid';
 import { UserSession } from '../api';
-import { IdeaGrid, UserGrid } from '.';
-import { Idea, Tag, User } from '../types';
+import { CheckboxTag, Idea, Tag } from '../types';
+import { NewIdeaFab } from '../components';
+import { IdeaGrid, InfoGrid } from '../grids';
 
 /**
- * UserLayout component styles
+ * HomeLayout component styles
  */
 const useStyles = makeStyles(
   (theme: Theme): Styles =>
@@ -15,16 +16,12 @@ const useStyles = makeStyles(
       root: {
         flexGrow: 1,
         display: 'flex',
-        justifyContent: 'center',
-        [theme.breakpoints.down('sm')]: {
-          flexDirection: 'column'
-        }
+        justifyContent: 'center'
       },
       grid: {
         width: '100%',
         marginTop: 72,
         [theme.breakpoints.down('sm')]: {
-          margin: '0 auto',
           marginTop: 66
         }
       }
@@ -32,50 +29,39 @@ const useStyles = makeStyles(
 );
 
 /**
- * UserLayout component prop types
+ * HomeLayout component prop types
  */
-interface UserLayoutProps {
+interface HomeLayoutProps {
   // Current user session
   user: UserSession;
-
-  // Current user being viewed
-  viewingUser: User | null;
-
   // Array of current ideas
   ideas: Idea[];
-
   // Array of all available tags
   allTags: Tag[];
+  // Array of currently-checked tag names
+  checkboxTags: CheckboxTag[];
 }
 
 /**
- * User page layout grid
+ * Home page layout grid
  */
-const UserLayout = ({
+const HomeLayout = ({
   user,
-  viewingUser,
   ideas,
-  allTags
-}: UserLayoutProps): React.ReactElement => {
+  allTags,
+  checkboxTags
+}: HomeLayoutProps): React.ReactElement => {
   const classes = useStyles();
 
   return (
     <div className={classes.root}>
       <Grid className={classes.grid} container spacing={6} justify="center">
-        <IdeaGrid
-          user={user}
-          self={
-            user.current && viewingUser
-              ? viewingUser.key === user.current.key
-              : false
-          }
-          ideas={ideas}
-          allTags={allTags}
-        />
-        <UserGrid user={user} viewingUser={viewingUser} />
+        <IdeaGrid user={user} self={true} ideas={ideas} allTags={allTags} />
+        <InfoGrid user={user} checkboxTags={checkboxTags} />
+        <NewIdeaFab user={user} allTags={allTags} />
       </Grid>
     </div>
   );
 };
 
-export default UserLayout;
+export default HomeLayout;
