+++ frontend/src/layouts/index.ts
@@ -1,0 +1,5 @@
+// Layouts barrel file
+export { default as Layout } from './Layout';
+export { default as HomeLayout } from './HomeLayout';
+export { default as IdeaLayout } from './IdeaLayout';
+export { default as UserLayout } from './UserLayout';
